The data is now shown properly.
Didn't change the JSON file names in variables. Some sample sizes ended in .5 so changed it from int to double. Forgot to add the display line of code for voter registration.

+++ Data/Model/ElectionDatam.swift
@@ -12,7 +12,7 @@ struct ElectionDatum: Codable//, Identifiable
     //var id = UUID().uuidString
     var state: String
     var candidateName: String
-    var sampleSize: Int
+    var sampleSize: Double
     var weight: Double
     var influence: Double
 }
